requirements.txt
kafka-python==2.0.2
confluent-kafka==2.0.2
facebook-sdk==3.1.0
requests==2.31.0
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
pandas==2.0.2
numpy==1.25.0
fake_useragent==1.2.0
loguru==0.6.0
python-dotenv==1.0.0
underthesea
